Refactor form data handling: move parse_formdata and related logic from router_helpers to form_data_parser
